requirements.txt
pickle5
numpy==1.23.1
tqdm==4.64.0
keras==2.9.0
tensorflow==2.9.1
nltk==3.7
Pillow==9.2.0
matplotlib==3.5.2
regex==2022.7.25
streamlit==1.11.1
requests==2.28.1
gTTS==2.2.4
googletrans==3.0.0
Flask==2.1.3